fixing: music on android
