+++ core/src/main/java/work/chiro/game/compatible/ResourceProvider.java
@@ -1,6 +1,10 @@
 package work.chiro.game.compatible;
 
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+
+import work.chiro.game.utils.Utils;
 
 public abstract class ResourceProvider {
     static private ResourceProvider instance = null;
@@ -16,5 +20,18 @@ public abstract class ResourceProvider {
 
     public abstract XImage<?> getImageFromResource(String path) throws IOException;
 
-    public abstract byte[] getSoundBytesFromResource(String path) throws IOException;
+    public byte[] getSoundBytesFromResource(String path) throws IOException {
+        InputStream fileInputStream = Utils.class.getResourceAsStream("/sounds/" + path);
+        if (fileInputStream == null) {
+            throw new IOException("file: " + path + " not found!");
+        }
+        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        final int maxSize = 1024;
+        int len;
+        byte[] b = new byte[maxSize];
+        while ((len = fileInputStream.read(b)) != -1) {
+            byteArrayOutputStream.write(b, 0, len);
+        }
+        return byteArrayOutputStream.toByteArray();
+    }
 }
